+++ MiDAS/R/MiDAS_stats.R
@@ -4,20 +4,20 @@
 #' calls using statistical model of choice.
 #'
 #' @inheritParams checkHlaCallsFormat
-#' @param model String specifying statistical model to use.
+#' @param model String specifying statistical model to use or corresponding
+#'   function.
 #' @param pheno Data frame holding phenotypic response variables.
 #' @param covar Data frame holding covariates.
 #' @param zygo Flag indicating whether zygosity should be added to
 #'   covariates. See details for further explanations.
 #' @param reduce_counts Flag indicating whether allele counts should be reduced
-#'   to presence / absence indicators.
-#' @param correction String specifying multiple testing correction method.
-#'
-#' Available choices for \code{model} include:
-#' \code{"coxph"} - Cox survival analysis,
-#' \code{"lm"} - Linear regression,
-#' \code{"glm.logit"} - Logistic regression,
-#' \code{"glm.nb"} - Negative binomial regression
+#'   to presence / absence indicators. See details for further explanations.
+#' @param correction String specifying multiple testing correction method. See
+#'   details for further information.
+#' @param exponentiate Logical indicating whether or not to exponentiate the the
+#'   coefficient estimates. This is typical for logistic and multinomial
+#'   regressions, but a bad idea if there is no log or logit link. Defaults to
+#'   FALSE.
 #'
 #' \code{pheno} and \code{covar} should be data frames with first column holding
 #' samples IDs and named \code{ID}. Those should correspond to \code{ID} column
@@ -25,7 +25,7 @@
 #'
 #' \code{zygo} indicate if additional covariate, indicating sample zygosity
 #' status, should be added to covariates. HLA allele counts for each sample
-#' can take following values \code{0, 1, 2}. To avoid implying ordering on those
+#' can take the following values \code{0, 1, 2}. To avoid implying ordering on those
 #' levels and effect size, this information can be split between two variables.
 #' If \code{zygo} is set to \code{TRUE} zygosity variable is added during model
 #' fitting, it specifies if sample is homozygous for an allele.
@@ -42,6 +42,7 @@
 #'   \code{hla_calls}.
 #'
 #' @examples
+#' library("survival")
 #' hla_calls_file <- system.file("extdata", "HLAHD_output_example.txt", package = "MiDAS")
 #' hla_calls <- readHlaCalls(hla_calls_file)
 #' pheno_file <- system.file("extdata", "pheno_example.txt", package = "MiDAS")
@@ -59,22 +60,12 @@
 #'                        correction = "BH"
 #' )
 #'
-#' # Logistic regression
-#' pheno <- pheno[, c(1, 3)]
-#' analyzeHlaAssociations(model = "glm.logit",
-#'                        hla_calls,
-#'                        pheno,
-#'                        covar,
-#'                        zygo = FALSE,
-#'                        reduce_counts = FALSE,
-#'                        correction = "BH"
-#' )
-#'
 #' @importFrom assertthat assert_that see_if is.flag is.string
 #' @importFrom broom tidy
 #' @importFrom dplyr left_join filter mutate rename
 #' @importFrom stats p.adjust
 #' @importFrom purrr map_dfr
+#'
 #' @export
 analyzeHlaAssociations <- function(model = "coxph",
                                    hla_calls,
@@ -82,11 +73,19 @@ analyzeHlaAssociations <- function(model = "coxph",
                                    covar,
                                    zygo = FALSE,
                                    reduce_counts = FALSE,
-                                   correction = "BH") {
+                                   correction = "BH",
+                                   exponentiate = FALSE) {
   assert_that(
-    is.string(model),
-    see_if(model %in% hlaAssocModels(),
-           msg = sprintf("model %s is not implemented", model)),
+    see_if(is.string(model) | is.function(model),
+           msg = "model have to be a string (a length one character vector) or a function"
+    ),
+    if (is.string(model)) {
+      see_if(is.function(get0(model)),
+             msg = sprintf("could not find function %s", model)
+      )
+    } else {
+      TRUE
+    },
     checkHlaCallsFormat(hla_calls),
     checkAdditionalData(pheno, hla_calls),
     checkAdditionalData(covar, hla_calls, accept.null = TRUE),
@@ -103,16 +102,24 @@ analyzeHlaAssociations <- function(model = "coxph",
   )
   alleles <- backquote(hla_data$alleles)
   response <- backquote(hla_data$response)
+  if (substitute(model) %in% c("coxph", "cph")) {
+    response <- paste0("Surv(", paste(response, collapse = ","), ")")
+  }
   covariate <- backquote(hla_data$covariate)
-  model_function <- hlaAssocModels(model = model,
-                                   response = response,
-                                   covariate = covariate,
-                                   data = hla_data$data
+  model_function <- hlaAssocModel(model = model,
+                                  response = response,
+                                  variable = covariate,
+                                  data = hla_data$data
   )
 
   results <- map_dfr(
     .x = alleles,
-    .f = ~tidy(model_function(.), exponentiate = FALSE) # TODO exponentiate could be passed somehow
+    .f = ~ tidy(updateModel(object = model_function,
+                            x = .,
+                            backquote = FALSE,
+                            collapse = " + "),
+                exponentiate = exponentiate
+    )
   )
   results <- mutate(results, term = gsub("`", "", term))
   results <- filter(results, checkAlleleFormat(term))
@@ -125,140 +132,19 @@ analyzeHlaAssociations <- function(model = "coxph",
 
 #' Association models for analysis of HLA alleles
 #'
-#' \code{hlaAssocModels} is a collection of pre-configured models for use with
-#' HLA alleles count table.
-#'
-#' \code{hlaAssocModels} is not intended to use by basic user.
+#' \code{hlaAssocModels} make creating statistical models for HLA association
+#' analyses easier.
 #'
 #' @inheritParams analyzeHlaAssociations
-#' @param response Character specifying response variables in \code{data}.
-#' @param covariate Character specifying covariates in \code{data}.
-#' @param data Data frame containing variables in the model.
+#' @param response String specifying response variable in \code{data}.
+#' @param variable Character specifying variables in \code{data}.
+#' @param data Data frame containing the variables in the model.
+#' @param ... Further arguments passed to \code{model} function.
 #'
-#' @return Function for fitting association model of choice, it takes allele
-#'   number as an argument. Additional arguments can be passed as well.
-#'
-#'   Returned function takes one required argument: HLA allele number
-#'   (\code{allele}). Additional parameters are passed to statistical model
-#'   function. Due to fact that allele numbers contains characters that have
-#'   special meanings in formulas, they should be backquoted. This can be
-#'   easily done with \link{backquote}. See examples section for general usage
-#'   case.
-#'
-#'   If model is equal to \code{NULL} names of available models are returned
-#'   instead.
+#' @return Fit from specified \code{model} function.
 #'
 #' @examples
-#' hla_calls_file <- system.file("extdata", "HLAHD_output_example.txt", package = "MiDAS")
-#' hla_calls <- readHlaCalls(hla_calls_file)
-#' hla_counts <- hlaCallsToCounts(hla_calls)
-#' pheno_file <- system.file("extdata", "pheno_example.txt", package = "MiDAS")
-#' pheno <- read.table(pheno_file, header = TRUE)
-#' covar_file <- system.file("extdata", "covar_example.txt", package = "MiDAS")
-#' covar <- read.table(covar_file, header = TRUE)
-#' data <- dplyr::left_join(hla_counts, pheno, by = "ID")
-#' data <- dplyr::left_join(data, covar, by = "ID")
-#' response <- paste(colnames(pheno[, -1]), collapse = ", ")
-#' covariate <- paste(colnames(covar[, -1]), collapse = " + ")
-#' fun <- hlaAssocModels(model = "coxph",
-#'                        response = response,
-#'                        covariate = covariate,
-#'                        data = data
-#' )
-#' allele <- backquote("A*01:01")
-#' fun(allele)
-#'
-#' @importFrom assertthat assert_that see_if
-#' @importFrom MASS glm.nb
-#' @importFrom stats as.formula binomial glm lm
-#' @importFrom survival coxph Surv
-#' @export
-hlaAssocModels <- function(model = NULL,
-                           response,
-                           covariate,
-                           data) {
-  if (is.null(model)) {
-    return(c("coxph", "lm", "glm.logit", "glm.nb"))
-  }
-  assert_that(
-    is.string(model),
-    is.character(response),
-    see_if(length(response) != 0, msg = "response can not be empty"),
-    see_if(is.character(covariate) | is.null(covariate),
-           msg = "covariate have to be a character or NULL"
-    ),
-    is.data.frame(data)
-  )
-
-  response_var <- paste(response, collapse = ", ")
-  if (length(covariate) != 0) {
-    covariate_var <- paste(covariate, collapse = " + ")
-  } else {
-    covariate_var <- "0"
-  }
-
-  model_function <- switch(
-    model,
-    coxph = function(allele, ...) { # Cox survival analysis
-      form <- sprintf("Surv(%s) ~ %s + %s", response_var, allele, covariate_var)
-      result <- coxph(formula = as.formula(form), data = data, ...)
-      return(result)
-    },
-    lm = function(allele, ...) { # Linear regression
-      form <- sprintf("%s ~ %s + %s", response_var, allele, covariate_var)
-      result <- lm(formula = as.formula(form), data = data, ...)
-      return(result)
-    },
-    glm.logit = function(allele, ...) { # Logistic regression
-      form <- sprintf("%s ~ %s + %s", response_var, allele, covariate_var)
-      result <- glm(
-        formula = as.formula(form),
-        family = binomial(link = "logit"),
-        data = data,
-        ...
-      )
-      return(result)
-    },
-    glm.nb = function(allele, ...) { # Negative binomial regression
-      form <- sprintf("%s ~ %s + %s", response_var, allele, covariate_var)
-      result <- glm.nb(formula = as.formula(form), data = data, ...)
-      return(result)
-    }
-  )
-  return(model_function)
-}
-
-#' Stepwise forward alleles subset selection
-#'
-#' \code{forwardAllelesSelection} does stepwise conditional testing adding the
-#' previous top-associated allele as covariate, until there’s no more
-#' significant alleles using a self-defined threshold.
-#'
-#' @param object object fitted by some model-fitting function.
-#' @param scope formula specifying a maximal model which should include the
-#'   current one. All additional terms in the maximal model with all marginal
-#'   terms in the original model are tried.
-#' @param th number specifying p-value threshold for a term to be included into
-#'   model.
-#' @param test string indicating test statistic to use for p-value calculation.
-#'   Can be either "F" or "Chisq".
-#' @param rss_th number specifying residual sum of squares threshold at which
-#'   function should stop adding additional terms.
-#' @param ... additional parameters passsed to \link[MASS]{addterm} function.
-#'
-#' All the variables in the \code{scope} should be defined in the \code{object}.
-#'
-#' The F test is only appropriate for lm and aov models, and perhaps for some
-#' over-dispersed glm models. The Chisq test can be an exact test (lm models
-#' with known scale) or a likelihood-ratio test depending on the method.
-#'
-#' As the residual sum of squares approaches \code{0} the perfect fit is
-#' obtained making further attempts at model selection nonsense, thus function
-#' is stopped. This behavior can be controlled using \code{rss_th}.
-#'
-#' @return selected model of the same class as \code{object} is returned.
-#'
-#' @examples
+#' library("survival")
 #' hla_calls_file <- system.file("extdata", "HLAHD_output_example.txt", package = "MiDAS")
 #' hla_calls <- readHlaCalls(hla_calls_file)
 #' pheno_file <- system.file("extdata", "pheno_example.txt", package = "MiDAS")
@@ -266,65 +152,228 @@ hlaAssocModels <- function(model = NULL,
 #' covar_file <- system.file("extdata", "covar_example.txt", package = "MiDAS")
 #' covar <- read.table(covar_file, header = TRUE)
 #' hla_data <- prepareHlaData(hla_calls, pheno, covar)
-#' object <- survival::coxph(survival::Surv(OS, OS_DIED) ~ AGE + SEX, data = hla_data$data)
-#' scope <- survival::Surv(OS, OS_DIED) ~ AGE + SEX + `B*57:01` + `C*07:02`
-#' forwardAllelesSelection(object, scope, th = 0.05, test = "Chisq")
+#' hlaAssocModel(model = "coxph",
+#'               response = "Surv(OS, OS_DIED)",
+#'               variable = c("AGE", "SEX"),
+#'               data = hla_data$data
+#' )
 #'
-#' @importFrom assertthat assert_that is.number is.string see_if
-#' @importFrom MASS addterm
+#' @importFrom assertthat assert_that is.string see_if
+#' @importFrom stats as.formula
 #' @importFrom purrr is_formula
+#'
+#' @export
+hlaAssocModel <- function(model,
+                          response,
+                          variable,
+                          data,
+                          ...) {
+  assert_that(
+    see_if(is.string(model) | is.function(model),
+           msg = "model have to be a string (a length one character vector) or a function"
+    ),
+    if (is.string(model)) {
+      see_if(is.function(get0(model)),
+             msg = sprintf("could not find function %s", model)
+      )
+    } else {
+      TRUE
+    },
+    see_if(is.string(response) | is_formula(response),
+           msg = "response have to be a string or formula"
+    ),
+    see_if(is.character(variable) | is_formula(variable),
+           msg = "variable have to be a character or formula"
+    ),
+    is.data.frame(data)
+  )
+
+  if (is.string(model)) {
+    model <- as.name(model)
+  }
+
+  if (is.character(response)) {
+    response <- paste0(response, " ~ 1")
+    response <- as.formula(response)
+  }
+
+  if (is.character(variable)) {
+    variable <- paste(variable, collapse = " + ")
+    variable <- paste(". ~ .", variable, sep = " + ")
+  }
+
+  form <- update(response, variable)
+  model_fun <- eval.parent(substitute(model(formula = form, data = data, ...)))
+
+  # Assert that model_fun is in fact model fit object
+  assert_that(
+    see_if(is.object(model_fun),
+           msg = sprintf("object returned by %s doesn't have OBJECT bit set",
+                         deparse(substitute(model))
+           )
+    ),
+    {
+      object_call <- get0("call", envir = as.environment(model_fun))
+      if (! is.null(object_call)) {
+        object_formula <- eval(substitute(formula, env = as.list(object_call)))
+        see_if(is_formula(object_formula),
+               msg = sprintf("object returned by %s is not a model with defined formula",
+                             deparse(substitute(model))
+               )
+        )
+      } else {
+        structure(FALSE,
+                  msg = sprintf("object returned by %s doesn't have an attribute 'call'",
+                                deparse(substitute(model))
+                  )
+        )
+      }
+    }
+  )
+
+  return(model_fun)
+}
+
+#' Stepwise conditional variables selection
+#'
+#' \code{stepwiseConditionalSelection} performs stepwise conditional testing
+#' adding the previous top-associated variable as covariate, until there’s no
+#' more significant variables based on a self-defined threshold.
+#'
+#' Selection criteria is the p-value from the test on coefficients values.
+#'
+#' @inheritParams checkHlaCallsFormat
+#' @inheritParams analyzeHlaAssociations
+#' @param th number specifying p-value threshold for a term to be included into
+#'   model.
+#' @param keep logical flag indicating if the output should be a list of models
+#'   resulting from each selection step. Default is to return only the final
+#'   model.
+#' @param rss_th number specifying residual sum of squares threshold at which
+#'   function should stop adding additional terms.
+#'
+#' As the residual sum of squares approaches \code{0} the perfect fit is
+#' obtained making further attempts at model selection nonsense, thus function
+#' is stopped. This behavior can be controlled using \code{rss_th}.
+#'
+#' @return selected model of the same class as \code{object} or list of models.
+#'   See \code{keep} parameter.
+#'
+#' @examples
+#' library("survival")
+#' hla_calls_file <- system.file("extdata", "HLAHD_output_example.txt", package = "MiDAS")
+#' hla_calls <- readHlaCalls(hla_calls_file)
+#' pheno_file <- system.file("extdata", "pheno_example.txt", package = "MiDAS")
+#' pheno <- read.table(pheno_file, header = TRUE)
+#' covar_file <- system.file("extdata", "covar_example.txt", package = "MiDAS")
+#' covar <- read.table(covar_file, header = TRUE)
+#' forwardConditionalSelection(model = "coxph",
+#'                             hla_calls = hla_calls,
+#'                             pheno = pheno,
+#'                             covar = covar,
+#'                             th = 0.05,
+#'                             keep = FALSE,
+#'                             rss_th = 1e-07
+#' )
+#'
+#' @importFrom assertthat assert_that is.flag is.number is.string see_if
+#' @importFrom MASS addterm
+#' @importFrom purrr is_formula map_dfr
 #' @importFrom rlang warn
 #' @importFrom stats formula resid update
-#' @importFrom stringi stri_startswith_fixed
 #' @export
-forwardAllelesSelection <- function(object,
-                                    scope,
-                                    th,
-                                    test = c("F", "Chisq"),
-                                    rss_th = 1e-07,
-                                    ...) {
+forwardConditionalSelection <- function(model,
+                                        hla_calls,
+                                        pheno,
+                                        covar,
+                                        th,
+                                        keep = FALSE,
+                                        rss_th = 1e-07) {
+
   assert_that(
-    see_if("formula" %in% attr(object, "names"),
-           msg = "object have to be a model"
+    see_if(is.string(model) | is.function(model),
+           msg = "model have to be a string (a length one character vector) or a function"
     ),
-    see_if(is_formula(scope), msg = "scope is not a formula"),
+    if (is.string(model)) {
+      see_if(is.function(get0(model)),
+             msg = sprintf("could not find function %s", model)
+      )
+    } else {
+      TRUE
+    },
+    checkHlaCallsFormat(hla_calls),
+    checkAdditionalData(pheno, hla_calls),
+    checkAdditionalData(covar, hla_calls, accept.null = TRUE),
     is.number(th),
-    is.string(test),
+    is.flag(keep),
     is.number(rss_th)
   )
 
-  test <- match.arg(test)
+  hla_data <- prepareHlaData(hla_calls, # Perhaps it would be beneficial to take this object outside, it has relatively simple structure so if someone needs to hack it it should be easy. Plus checking the data you are putting into functions would be beneficial.
+                             pheno,
+                             covar,
+                             zygo = FALSE,
+                             reduce_counts = FALSE
+  )
 
-  criterium_pattern <- "Pr("
-  scope_vars <- all.vars(scope)
-  cur_vars <- all.vars(object$call$formula)
+  alleles <- hla_data$alleles
+  response <- backquote(hla_data$response)
+  if (substitute(model) %in% c("coxph", "cph")) {
+    response <- paste0("Surv(", paste(response, collapse = ","), ")")
+  }
+  covariate <- backquote(hla_data$covariate)
+  object <- hlaAssocModel(model = model,
+                          response = response,
+                          variable = covariate,
+                          data = hla_data$data
+  )
 
-  while (! all(scope_vars %in% cur_vars)) {
-    temp <- addterm(object, scope, test = test, ...)
-    criterium_column <- stri_startswith_fixed(colnames(temp), criterium_pattern)
-    assert_that(
-      see_if(sum(criterium_column) == 1,
-             msg = "Test criterium column couldn't be found..."
-      )
+  vars <- alleles
+  prev_formula <- formula(object)
+  prev_vars <- all.vars(prev_formula)
+
+  best <- list(object)
+  i <- 2
+
+  while (TRUE) {
+    new_vars <- vars[! vars %in% prev_vars]
+
+    results <- map_dfr(
+      .x = new_vars,
+      .f = ~ tidy(updateModel(object = object,
+                              x = .,
+                              backquote = TRUE,
+                              collapse = " + "
+      ))
     )
-    test_val <- temp[-1, criterium_column]
-    i_min <- which.min(test_val)
-    if (length(i_min) == 0) break()
-    if (test_val[i_min] > th) break()
-    new_var <- rownames(temp)[i_min + 1]
-    new_formula <- formula(object)
-    new_formula <- as.formula(
-      paste(new_formula[2], "~", paste(new_formula[3], new_var, sep = " + "))
+    results <- results[results$term %in% backquote(new_vars), ]
+    results <- results[! is.infinite(results$p.value), ]
+
+    i_min <- which.min(results$p.value)
+    if (length(i_min) == 0) break
+    if (results$p.value[i_min] > th) break
+
+    object <- updateModel(object,
+                          results$term[i_min],
+                          backquote = FALSE,
+                          collapse = " + "
     )
-    object <- update(object, new_formula)
+
     if (sum(resid(object) ^ 2) <= rss_th) {
       warn("Perfect fit was reached attempting further model selection is nonsense.")
-      break()
+      break
     }
-    cur_vars <- all.vars(object$call$formula)
+    prev_formula <- formula(object)
+    prev_vars <- all.vars(prev_formula)
+    best[[i]] <- object
+    i <- i + 1
   }
 
-  return(object)
+  if (! keep) {
+    best <- best[[length(best)]]
+  }
+
+  return(best)
 }
 
 #' Prepare data for statistical analysis
@@ -345,6 +394,9 @@ forwardAllelesSelection <- function(object,
 #'                zygo = FALSE,
 #'                reduce_counts = FALSE
 #' )
+#'
+#' @importFrom assertthat assert_that is.flag see_if
+#'
 #' @export
 prepareHlaData <- function(hla_calls,
                            pheno,