+++ MiDAS/R/MiDAS_utils.R
@@ -255,7 +255,7 @@ checkHlaCallsFormat <- function(hla_calls) {
     see_if(! any(vapply(hla_calls, is.factor, logical(length = 1))),
            msg = "hla_calls can't contain factors"
     ),
-    see_if(! all(checkAlleleFormat(hla_calls[, 1]), na.rm = TRUE),
+    see_if(! all(checkAlleleFormat(as.character(hla_calls[, 1])), na.rm = TRUE),
            msg = "first column of hla_calls should specify samples id"
     ),
     see_if(all(checkAlleleFormat(unlist(hla_calls[, -1])), na.rm = TRUE),
@@ -284,6 +284,7 @@ checkHlaCallsFormat <- function(hla_calls) {
 #' @export
 backquote <- function(x) {
   assert_that(is.character(x))
+  x <- gsub("`", "", x)
   backquted <- paste0("`", x, "`")
   return(backquted)
 }
@@ -344,4 +345,73 @@ checkAdditionalData <- function(data_frame,
   }
 
   return(TRUE)
+}
+
+#' Add new variables to statistical model
+#'
+#' \code{updateModel} will add new variables to model and re-fit it.
+#'
+#' @param object An existing fit from a model function such as lm, glm and many
+#'   others.
+#' @param x Character vector specifying variables to be added to model or a
+#'   formula giving a template which specifies how to update.
+#' @param backquote Logical indicating if added variables should be quoted.
+#'   Longer than one element vectors are accepted as well, specifying which new
+#'   variables should be backquoted. Only relevant if x is of type character.
+#' @param collapse Character specifying how new characters should be added to
+#'   old formula. Only relevant if x is of type character.
+#'
+#' @return Updated fit of input model.
+#'
+#' @importFrom assertthat assert_that
+#' @importFrom stats update
+#' @importFrom purrr is_formula
+#'
+#' @examples
+#' library("survival")
+#' hla_calls_file <- system.file("extdata", "HLAHD_output_example.txt", package = "MiDAS")
+#' hla_calls <- readHlaCalls(hla_calls_file)
+#' pheno_file <- system.file("extdata", "pheno_example.txt", package = "MiDAS")
+#' pheno <- read.table(pheno_file, header = TRUE)
+#' covar_file <- system.file("extdata", "covar_example.txt", package = "MiDAS")
+#' covar <- read.table(covar_file, header = TRUE)
+#' hla_data <- prepareHlaData(hla_calls, pheno, covar)
+#' coxmod <- hlaAssocModel(model = "coxph",
+#'                         response = "Surv(OS, OS_DIED)",
+#'                         variable = "1",
+#'                         data = hla_data$data
+#' )
+#' updateModel(coxmod, "A*01:01", backquote = TRUE, collapse = " + ")
+#'
+#' @export
+updateModel <- function(object, x, backquote = TRUE, collapse = " + ") {
+  assert_that( # This could be simplified to simple object type is wrong / not supported
+    see_if(is.object(object),
+           msg = "object have to have the internal OBJECT bit set"
+    ),
+    {
+      object_call <- get0("call", envir = as.environment(object))
+      if (! is.null(object_call)) {
+        object_formula <- eval(substitute(formula, env = as.list(object_call)))
+        see_if(is_formula(object_formula),
+               msg = "object have to be a model with defined formula"
+        )
+      } else {
+        structure(FALSE, msg = "object have to have an attribute 'call'")
+      }
+    },
+    see_if(is.character(x) | is_formula(x),
+           msg = "x have to be a string (a length one character vector) or formula"
+    )
+  )
+
+  if (is.character(x)) {
+    x[backquote] <- backquote(x[backquote])
+    x <- paste0(". ~ . + ", paste(x, collapse = collapse))
+  }
+
+  new_object <- update(object = object, x, evaluate = FALSE)
+  new_object <- eval.parent(new_object)
+
+  return(new_object)
 }